refactor classes into their own files
